requirements.txt
streamlit>=1.37.0
langchain-core>=0.2.35
langchain-community>=0.2.14
langchain-groq>=0.1.3
python-dotenv>=1.0.1
pydantic>=2.8.2
pypdf>=4.2.0
python-docx>=1.1.2
chromadb>=0.5.5
pandas>=2.2.2
beautifulsoup4>=4.12.3
lxml>=5.2.1
requests>=2.32.3
email-validator>=2.1.0.post1
SQLAlchemy>=2.0.32
passlib[bcrypt]>=1.7.4
streamlit-oauth>=0.1.6
google-auth>=2.33.0
google-auth-oauthlib>=1.2.1